Remove unused files and new suffix
