+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/ReaderDiscoverAdapter.kt
@@ -9,7 +9,7 @@ import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderPostNewU
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderPostUiState
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderRecommendedBlogsCardUiState
 import org.wordpress.android.ui.reader.discover.viewholders.ReaderAnnouncementCardViewHolder
-import org.wordpress.android.ui.reader.discover.viewholders.ReaderInterestsCardNewViewHolder
+import org.wordpress.android.ui.reader.discover.viewholders.ReaderInterestsCardViewHolder
 import org.wordpress.android.ui.reader.discover.viewholders.ReaderPostNewViewHolder
 import org.wordpress.android.ui.reader.discover.viewholders.ReaderPostViewHolder
 import org.wordpress.android.ui.reader.discover.viewholders.ReaderRecommendedBlogsCardNewViewHolder
@@ -44,7 +44,7 @@ class ReaderDiscoverAdapter(
                 parent
             )
 
-            INTEREST_VIEW_TYPE -> ReaderInterestsCardNewViewHolder(uiHelpers, parent)
+            INTEREST_VIEW_TYPE -> ReaderInterestsCardViewHolder(uiHelpers, parent)
 
             RECOMMENDED_BLOGS_VIEW_TYPE -> ReaderRecommendedBlogsCardNewViewHolder(parent, imageManager)
