+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/ReaderRecommendedBlogsAdapter.kt
@@ -5,15 +5,13 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderRecommendedBlogsCardUiState.ReaderRecommendedBlogUiState
 import org.wordpress.android.ui.reader.discover.viewholders.ReaderRecommendedBlogViewHolder
-import org.wordpress.android.ui.utils.UiHelpers
 import org.wordpress.android.util.image.ImageManager
 
 class ReaderRecommendedBlogsAdapter(
     private val imageManager: ImageManager,
-    private val uiHelpers: UiHelpers,
 ) : ListAdapter<ReaderRecommendedBlogUiState, ReaderRecommendedBlogViewHolder>(RecommendedBlogsDiffUtil()) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ReaderRecommendedBlogViewHolder {
-        return ReaderRecommendedBlogViewHolder(parent, imageManager, uiHelpers)
+        return ReaderRecommendedBlogViewHolder(parent, imageManager)
     }
 
     override fun onBindViewHolder(holder: ReaderRecommendedBlogViewHolder, position: Int) {
