+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/viewholders/ReaderInterestViewHolder.kt
@@ -1,7 +1,6 @@
 package org.wordpress.android.ui.reader.discover.viewholders
 
 import android.view.ViewGroup
-import androidx.appcompat.content.res.AppCompatResources.getColorStateList
 import androidx.recyclerview.widget.RecyclerView
 import org.wordpress.android.databinding.ReaderInterestItemBinding
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderInterestsCardUiState.ReaderInterestUiState
@@ -14,13 +13,7 @@ class ReaderInterestViewHolder(
     private val binding: ReaderInterestItemBinding = parent.viewBinding(ReaderInterestItemBinding::inflate)
 ) : RecyclerView.ViewHolder(binding.root) {
     fun onBind(uiState: ReaderInterestUiState) = with(binding) {
-        uiHelpers.setTextOrHide(interest, uiState.interest)
-        interest.setOnClickListener { uiState.onClicked.invoke(uiState.interest) }
-
-        with(uiState.chipStyle) {
-            interest.setChipStrokeColorResource(chipStrokeColorResId)
-            interest.setChipBackgroundColorResource(chipFillColorResId)
-            interest.setTextColor(getColorStateList(interest.context, chipFontColorResId))
-        }
+        uiHelpers.setTextOrHide(chip, uiState.interest)
+        chip.setOnClickListener { uiState.onClicked.invoke(uiState.interest) }
     }
 }
