+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/viewholders/ReaderRecommendedBlogViewHolder.kt
@@ -4,7 +4,6 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import org.wordpress.android.databinding.ReaderRecommendedBlogItemBinding
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderRecommendedBlogsCardUiState.ReaderRecommendedBlogUiState
-import org.wordpress.android.ui.utils.UiHelpers
 import org.wordpress.android.util.extensions.viewBinding
 import org.wordpress.android.util.image.ImageManager
 import org.wordpress.android.util.image.ImageType.BLAVATAR_CIRCULAR
@@ -12,25 +11,30 @@ import org.wordpress.android.util.image.ImageType.BLAVATAR_CIRCULAR
 class ReaderRecommendedBlogViewHolder(
     parent: ViewGroup,
     private val imageManager: ImageManager,
-    private val uiHelpers: UiHelpers,
     private val binding: ReaderRecommendedBlogItemBinding =
         parent.viewBinding(ReaderRecommendedBlogItemBinding::inflate)
 ) : RecyclerView.ViewHolder(binding.root) {
-    fun onBind(uiState: ReaderRecommendedBlogUiState) = with(binding) {
-        with(uiState) {
-            siteName.text = name
-            siteUrl.text = url
-            uiHelpers.setTextOrHide(siteDescription, description)
-            siteFollowButton.apply {
-                setIsFollowed(isFollowed)
-                contentDescription = context.getString(followContentDescription.stringRes)
-                setOnClickListener {
-                    onFollowClicked(uiState)
-                }
-            }
-            updateBlogImage(iconUrl)
+    fun onBind(uiState: ReaderRecommendedBlogUiState) =
+        with(binding) {
+            siteName.text = uiState.name
+            siteUrl.text = uiState.url
+            updateSiteFollowButton(uiState, this)
+            updateBlogImage(uiState.iconUrl)
             root.setOnClickListener {
-                onItemClicked(blogId, feedId, isFollowed)
+                uiState.onItemClicked(uiState.blogId, uiState.feedId, uiState.isFollowed)
+            }
+        }
+
+    private fun updateSiteFollowButton(
+        uiState: ReaderRecommendedBlogUiState,
+        binding: ReaderRecommendedBlogItemBinding
+    ) {
+        with(binding.siteFollowButton) {
+            isEnabled = uiState.isFollowEnabled
+            setIsFollowed(uiState.isFollowed)
+            contentDescription = context.getString(uiState.followContentDescription.stringRes)
+            setOnClickListener {
+                uiState.onFollowClicked(uiState)
             }
         }
     }
