+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/viewholders/ReaderRecommendedBlogsCardNewViewHolder.kt
@@ -1,21 +1,21 @@
 package org.wordpress.android.ui.reader.discover.viewholders
 
 import android.view.ViewGroup
-import org.wordpress.android.databinding.ReaderRecommendedBlogsCardNewBinding
+import org.wordpress.android.databinding.ReaderRecommendedBlogsCardBinding
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderRecommendedBlogsCardUiState
-import org.wordpress.android.ui.reader.discover.ReaderRecommendedBlogsNewAdapter
+import org.wordpress.android.ui.reader.discover.ReaderRecommendedBlogsAdapter
 import org.wordpress.android.util.extensions.viewBinding
 import org.wordpress.android.util.image.ImageManager
 
 class ReaderRecommendedBlogsCardNewViewHolder(
     parentView: ViewGroup,
     imageManager: ImageManager,
-) : ReaderViewHolder<ReaderRecommendedBlogsCardNewBinding>(
-    parentView.viewBinding(ReaderRecommendedBlogsCardNewBinding::inflate)
+) : ReaderViewHolder<ReaderRecommendedBlogsCardBinding>(
+    parentView.viewBinding(ReaderRecommendedBlogsCardBinding::inflate)
 ) {
     private val recommendedBlogsAdapter =
-        ReaderRecommendedBlogsNewAdapter(imageManager)
+        ReaderRecommendedBlogsAdapter(imageManager)
 
     init {
         with(binding) {
