+++ WordPress/src/main/res/layout/reader_cardview_xpost.xml
