+++ WordPress/src/main/res/layout/reader_interest_item.xml
@@ -1,23 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<com.google.android.material.chip.Chip xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/reader_interest_container"
+    android:id="@+id/chip"
+    style="@style/ReaderExpandableTagsViewChipInterestNew"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:layout_marginStart="@dimen/margin_small"
-    android:layout_marginEnd="@dimen/margin_small"
-    android:layout_marginBottom="@dimen/margin_extra_large"
-    android:layout_marginTop="@dimen/margin_extra_large">
-
-    <com.google.android.material.chip.Chip
-        android:id="@+id/interest"
-        style="@style/ReaderTextView.Interests.ListItem.Chip"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:chipStrokeWidth="@dimen/divider_size"
-        app:chipStartPadding="@dimen/margin_large"
-        app:chipEndPadding="@dimen/margin_large"
-        tools:text="Photography"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
+    tools:text="A Text in a chip" />
