--- WordPress/src/main/res/layout/reader_interest_item_new.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.chip.Chip xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/chip"
-    style="@style/ReaderExpandableTagsViewChipInterestNew"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    tools:text="A Text in a chip" />
