--- WordPress/src/main/res/layout/reader_recommended_blog_item_new.xml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="?attr/selectableItemBackground"
-    android:paddingBottom="@dimen/margin_medium_large"
-    android:paddingEnd="@dimen/margin_extra_large"
-    android:paddingStart="@dimen/margin_extra_large"
-    android:paddingTop="@dimen/margin_medium_large">
-
-    <androidx.appcompat.widget.AppCompatImageView
-        android:id="@+id/site_icon"
-        style="@style/ReaderImageView.Avatar"
-        android:background="@drawable/bg_oval_placeholder"
-        android:contentDescription="@null"
-        android:importantForAccessibility="no"
-        android:src="@drawable/bg_oval_placeholder_image_32dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/site_name"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="@dimen/margin_large"
-        android:layout_marginStart="@dimen/margin_large"
-        android:ellipsize="end"
-        android:fontFamily="sans-serif-medium"
-        android:maxLines="1"
-        android:textAppearance="?attr/textAppearanceLabelLarge"
-        android:textColor="?attr/colorOnSurface"
-        android:textSize="@dimen/text_sz_medium"
-        app:layout_constraintBottom_toTopOf="@+id/site_url"
-        app:layout_constraintEnd_toStartOf="@id/site_follow_button"
-        app:layout_constraintStart_toEndOf="@id/site_icon"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed"
-        tools:text="Site site Site site Site site Site site Site site Site site Site site " />
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/site_url"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:textAppearance="?attr/textAppearanceBodyMedium"
-        android:textColor="@color/reader_improvements_recommended_section_text"
-        android:textSize="@dimen/text_sz_small"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="@id/site_name"
-        app:layout_constraintStart_toStartOf="@id/site_name"
-        app:layout_constraintTop_toBottomOf="@+id/site_name"
-        tools:text="site.com site.com site.com site.com site.com site.com site.com site.com " />
-
-    <org.wordpress.android.ui.reader.views.ReaderFollowButton
-        android:id="@+id/site_follow_button"
-        style="@style/Reader.Follow.Button.New"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/reader_btn_subscribe"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
