+++ WordPress/src/main/res/layout/reader_recommended_blogs_card.xml
@@ -32,7 +32,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/recommended_blogs_header"
                 tools:itemCount="3"
-                tools:listitem="@layout/reader_recommended_blog_item_new" />
+                tools:listitem="@layout/reader_recommended_blog_item" />
 
             <com.google.android.material.textview.MaterialTextView
                 android:id="@+id/recommended_blogs_header"
